creates a file with mode read and write or create

+++ 0x15-file_io/1-create_file.c
@@ -10,7 +10,7 @@
 int create_file(const char *filename,
 		__attribute__((unused))char *text_content)
 {
-	size_t op;
+	int op;
 
 	op = open(filename, O_RDWR | O_CREAT);
